fix: address review feedback on backup/upgrade/rollback PR
- Bind auto-rollback restore to the specific backup from the current
  upgrade attempt instead of scanning for the latest pre-upgrade backup
  on disk. This prevents restoring stale data from a previous upgrade
  cycle when the current backup creation failed.

- Add 401-retry logic to the standalone `vellum backup` command for
  consistency with the shared `createBackup()` in backup-ops.ts.

- Extract broadcastUpgradeEvent, captureContainerEnv, and waitForReady
  from upgrade.ts into cli/src/lib/upgrade-lifecycle.ts, fixing the
  cross-command import violation (rollback.ts was importing from
  ./upgrade, violating cli/CONTRIBUTING.md module boundary rules).
