+++ cli/src/commands/upgrade.ts
@@ -13,7 +13,6 @@ import type { AssistantEntry } from "../lib/assistant-config";
 import {
   captureImageRefs,
   clearSigningKeyBootstrapLock,
-  DOCKER_READY_TIMEOUT_MS,
   GATEWAY_INTERNAL_PORT,
   dockerResourceNames,
   migrateCesSecurityFiles,
@@ -30,16 +29,19 @@ import {
 import {
   loadBootstrapSecret,
   saveBootstrapSecret,
-  loadGuardianToken,
 } from "../lib/guardian-token";
 import {
   createBackup,
-  findLatestPreUpgradeBackup,
   pruneOldBackups,
   restoreBackup,
 } from "../lib/backup-ops.js";
 import { emitCliError, categorizeUpgradeError } from "../lib/cli-error.js";
-import { exec, execOutput } from "../lib/step-runner";
+import { exec } from "../lib/step-runner";
+import {
+  broadcastUpgradeEvent,
+  captureContainerEnv,
+  waitForReady,
+} from "../lib/upgrade-lifecycle.js";
 import { commitWorkspaceState } from "../lib/workspace-git.js";
 
 interface UpgradeArgs {
@@ -154,101 +156,6 @@ function resolveTargetAssistant(nameArg: string | null): AssistantEntry {
     emitCliError("ASSISTANT_NOT_FOUND", msg);
   }
   process.exit(1);
-}
-
-/**
- * Capture environment variables from a running Docker container so they
- * can be replayed onto the replacement container after upgrade.
- */
-export async function captureContainerEnv(
-  containerName: string,
-): Promise<Record<string, string>> {
-  const captured: Record<string, string> = {};
-  try {
-    const raw = await execOutput("docker", [
-      "inspect",
-      "--format",
-      "{{json .Config.Env}}",
-      containerName,
-    ]);
-    const entries = JSON.parse(raw) as string[];
-    for (const entry of entries) {
-      const eqIdx = entry.indexOf("=");
-      if (eqIdx > 0) {
-        captured[entry.slice(0, eqIdx)] = entry.slice(eqIdx + 1);
-      }
-    }
-  } catch {
-    // Container may not exist or not be inspectable
-  }
-  return captured;
-}
-
-/**
- * Poll the gateway `/readyz` endpoint until it returns 200 or the timeout
- * elapses. Returns whether the assistant became ready.
- */
-export async function waitForReady(runtimeUrl: string): Promise<boolean> {
-  const readyUrl = `${runtimeUrl}/readyz`;
-  const start = Date.now();
-
-  while (Date.now() - start < DOCKER_READY_TIMEOUT_MS) {
-    try {
-      const resp = await fetch(readyUrl, {
-        signal: AbortSignal.timeout(5000),
-      });
-      if (resp.ok) {
-        const elapsedSec = ((Date.now() - start) / 1000).toFixed(1);
-        console.log(`Assistant ready after ${elapsedSec}s`);
-        return true;
-      }
-      let detail = "";
-      try {
-        const body = await resp.text();
-        const json = JSON.parse(body);
-        const parts = [json.status];
-        if (json.upstream != null) parts.push(`upstream=${json.upstream}`);
-        detail = ` — ${parts.join(", ")}`;
-      } catch {
-        // ignore parse errors
-      }
-      console.log(`Readiness check: ${resp.status}${detail} (retrying...)`);
-    } catch {
-      // Connection refused / timeout — not up yet
-    }
-    await new Promise((r) => setTimeout(r, 1000));
-  }
-
-  return false;
-}
-
-/**
- * Best-effort broadcast of an upgrade lifecycle event to connected clients
- * via the gateway's upgrade-broadcast proxy. Uses guardian token auth.
- * Failures are logged but never block the upgrade flow.
- */
-export async function broadcastUpgradeEvent(
-  gatewayUrl: string,
-  assistantId: string,
-  event: Record<string, unknown>,
-): Promise<void> {
-  try {
-    const token = loadGuardianToken(assistantId);
-    const headers: Record<string, string> = {
-      "Content-Type": "application/json",
-    };
-    if (token?.accessToken) {
-      headers["Authorization"] = `Bearer ${token.accessToken}`;
-    }
-    await fetch(`${gatewayUrl}/v1/admin/upgrade-broadcast`, {
-      method: "POST",
-      headers,
-      body: JSON.stringify(event),
-      signal: AbortSignal.timeout(3000),
-    });
-  } catch {
-    // Best-effort — gateway/daemon may already be shutting down or not yet ready
-  }
 }
 
 async function upgradeDocker(
@@ -530,15 +437,17 @@ async function upgradeDocker(
 
         const rollbackReady = await waitForReady(entry.runtimeUrl);
         if (rollbackReady) {
-          // Restore data from pre-upgrade backup if available
-          const latestBackup = findLatestPreUpgradeBackup(entry.assistantId);
-          if (latestBackup) {
+          // Restore data from the backup created for THIS upgrade attempt.
+          // Only use the specific backupPath — never scan for the latest
+          // backup on disk, which could be from a previous upgrade cycle
+          // and contain stale data.
+          if (backupPath) {
             console.log(`📦 Restoring data from pre-upgrade backup...`);
-            console.log(`   Source: ${latestBackup}`);
+            console.log(`   Source: ${backupPath}`);
             const restored = await restoreBackup(
               entry.runtimeUrl,
               entry.assistantId,
-              latestBackup,
+              backupPath,
             );
             if (restored) {
               console.log("   ✅ Data restored successfully\n");
@@ -549,7 +458,7 @@ async function upgradeDocker(
             }
           } else {
             console.log(
-              "ℹ️  No pre-upgrade backup found, skipping data restoration\n",
+              "ℹ️  No pre-upgrade backup was created for this attempt, skipping data restoration\n",
             );
           }
